tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/empty_keyframes_rule.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - empty @keyframes rule</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#keyframes">
   <meta name="flags" content="dom">
   <meta name="assert" content="@keyframes with no keyframes is valid">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes test {
         from {
            left: 100px;
         }
         to {
            left: 0px;
         }
      }
      @keyframes test {}
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
      }
      .test {
         animation: test 1s;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         div.className = 'test';
         var passed = getComputedStyle(div, null).left == 'auto';
         test(function(){assert_true(passed);}, testName); done();
         log.data = passed ? 'PASS if the square did not move' : 'FAIL';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The square below should not move, and it should not say FAIL below.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
